Comments added to the code

+++ MiniStat.h
@@ -99,8 +99,8 @@ public:
     void runCV(uint8_t user_gain, uint8_t cycles, uint16_t scan_rate);
     void runACV(uint8_t user_gain, uint8_t cycles, uint16_t scan_rate, uint16_t startV, uint16_t endV, uint16_t amplitude, uint16_t freq);
     void runPulseV(uint8_t user_gain, uint8_t cycles, uint16_t frequency, uint16_t pulse_width, int pulse_amplitude, uint8_t pulse_per_cycle);
-	void runSWV(uint8_t user_gain, uint8_t cycles, uint16_t startV, uint16_t endV,  int pulse_amp, uint16_t volt_step, uint16_t pulse_freq);
-	void runDPV(uint8_t user_gain, uint8_t cycles, uint16_t startV, uint16_t endV, int step_size, int pulse_amp, uint16_t sample_period, uint16_t pulse_freq);
+	void runSWV(uint8_t user_gain, uint8_t cycles, int startV, int endV,  int pulse_amp, int volt_step, uint16_t pulse_freq);
+	void runDPV(uint8_t user_gain, uint8_t cycles, int startV, int endV, int step_size, int pulse_amp, uint16_t sample_period, uint16_t pulse_freq);
 	void method(uint8_t bias, uint16_t scan_rate, int polarity);
 	void runAMP(uint16_t user_gain, int voltage, uint16_t time, int samples);
     void save();
